+++ UnstoppableWallet/UnstoppableWallet/Modules/AmountData.swift
