+++ UnstoppableWallet/UnstoppableWallet/Modules/Send/Settings/TimeLock/TimeLockViewModel.swift
@@ -2,6 +2,7 @@ import RxSwift
 import RxRelay
 import RxCocoa
 import Foundation
+import Hodler
 
 class TimeLockViewModel {
     private let disposeBag = DisposeBag()
@@ -55,6 +56,23 @@ extension TimeLockViewModel: IDropDownListViewModel {
 
     var selectedItemDriver: Driver<String?> {
         valueRelay.asDriver()
+    }
+
+}
+
+extension HodlerPlugin.LockTimeInterval {
+
+    static func title(lockTimeInterval: HodlerPlugin.LockTimeInterval?) -> String {
+        guard let lockTimeInterval = lockTimeInterval else {
+            return "send.hodler_locktime_off".localized
+        }
+
+        switch lockTimeInterval {
+        case .hour: return "send.hodler_locktime_hour".localized
+        case .month: return "send.hodler_locktime_month".localized
+        case .halfYear: return "send.hodler_locktime_half_year".localized
+        case .year: return "send.hodler_locktime_year".localized
+        }
     }
 
 }
